% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SegmentedCellExperiment-package.R
\docType{package}
\name{SegmentedCellExperiment-package}
\alias{SegmentedCellExperiment-package}
\title{\code{SegmentedCellExperiment} package}
\description{
S4 Class for segmented spatial 'omics data
}
\details{
Development version can be found on 
\href{https://github.com/ellispatrick/SegmentedCellExperiment}{GitHub}
}
