% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R,
%   R/SegmentedCellExperiment-class.R, R/SegmentedCellExperiment-package.R
\docType{class}
\name{SegmentedCellExperiment-class}
\alias{SegmentedCellExperiment-class}
\alias{SegmentedCellExperiment}
\alias{SegmentedCellExperiment,SegmentedCellExperiment-method}
\title{The SegmentedCellExperiment class}
\usage{
SegmentedCellExperiment(
  cellData,
  cellProfiler = FALSE,
  spatialCoords = NULL,
  cellTypeString = NULL,
  intensityString = NULL,
  morphologyString = NULL,
  cellIDString = NULL,
  imageCellIDString = NULL,
  imageIDString = NULL
)
}
\arguments{
\item{cellData}{A data frame that contains at least the columns x and y giving the location of each cell.}

\item{cellProfiler}{A logical indicating that cellData is in a format similar to what cellProfiler outputs.}

\item{spatialCoords}{The column names corresponding to spatial coordinates. eg. x, y, z...}

\item{cellTypeString}{The name of the column that contains cell type calls.}

\item{intensityString}{A string which can be used to identify the columns which contain marker intensities. (This needs to be extended to take the column names themselves.)}

\item{morphologyString}{A string which can be used to identify the columns which contains morphology information.}

\item{cellIDString}{The column name for cellID.}

\item{imageCellIDString}{The column name for imageCellID.}

\item{imageIDString}{The column name for imageIDString}
}
\value{
A SegmentedCellExperiment object
}
\description{
S4 Class for segmented spatial 'omics data
}
\details{
See the README on
\href{https://cran.r-project.org/package=googlesheets/README.html}{CRAN}
or \href{https://github.com/jennybc/googlesheets#readme}{GitHub}
}
\examples{
### Something that resembles cellProfiler data

set.seed(51773)

n = 10

cells <- data.frame(row.names = seq_len(n))
cells$ObjectNumber <- seq_len(n)
cells$ImageNumber <- rep(1:2,c(n/2,n/2))
cells$AreaShape_Center_X <- runif(n)
cells$AreaShape_Center_Y <- runif(n)
cells$AreaShape_round <- rexp(n)
cells$AreaShape_diameter <- rexp(n, 2)
cells$Intensity_Mean_CD8 <- rexp(n, 10)
cells$Intensity_Mean_CD4 <- rexp(n, 10)

cellExp <- SegmentedCellExperiment(cells, cellProfiler = TRUE)

### Cluster cell types
intensities <- intensity(cellExp)
kM <- kmeans(intensities,2)
cellType(cellExp) <- paste('cluster',kM$cluster, sep = '')
location(cellExp)

}
