% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SegmentedCellExperiment-package.R
\docType{package}
\name{SegmentedCellExperiment-package}
\alias{SegmentedCellExperiment-package}
\title{\code{SegmentedCellExperiment} package}
\description{
This package describes an S4 class for storing data from segmented 
imaging cytometry and spatial omics data. It extends DataFrame and defines 
methods that take advantage of DataFrame nesting to represent elements of 
cell-based experiments with spatial orientation that are commonly 
encountered. This object is able to store information on a cell's spatial 
location, cellType, morphology, intensity of gene/protein marks as well as 
image level phenotype information.
}
\details{
Development version can be found on 
\href{https://github.com/ellispatrick/SegmentedCellExperiment}{GitHub}
}
